--- src/components/EditScore/ScoreInputForm.js
@@ -1,9 +1,0 @@
-// import React, { Component } from "react";
-// import SingleLineScoreInput  from "./SingleLineScoreInput";
-
-
-// class ScoreInputForm extends Component {
-
-// }
-
-// export default ScoreInputForm;
